zephyr: CMakeLists.txt hack
It seems that when compiling with platformio, when a Kconfig is set in
the zephyr folder, it overrides Zephyr's Kconfig which causes a bunch
of defines to be missing. It seems that the only way to get the Kconfig
to work when building with West and have Platformio compatibility, is
to mask the Kconfig file during the Platformio build.

- Also fixes erroneously having -DZEPHYR flag on private scope by
  setting it to public scope