threading: Add new primitive RefGuard
RefGuard is a reference counter that allows graceful destruction by
allowing user to enter/leave with failure if the reference counter was
shutdown by remote.

It is useful for coordinated shutdowns where one endpoint wants to
shutdown but others are still relying on the object staying alive.

By entering/leaving in a critical section, it can be guaranteed that
both sides can gracefully track state and no UAF occurs.